# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Aviary API
  version: 0.1.0
  summary: Taxonomy + entity graph service with path-based data resolution.
  description: |
    Aviary is a schema-flexible entity store. Users declare **taxonomies**
    (schemas) at runtime, each with typed **fields** and **relationships** to
    other taxonomies. **Entities** are records whose `attributes` are a flat
    map of scalar values keyed by their taxonomy's field keys.

    Relationships are **computed, not stored**: `to_one` / `to_many` resolve
    at read time by matching `source_field` values on one side to
    `target_field` values on the other; `to_many_through` composes an
    ordered chain of other relationships. There are no foreign keys between
    entities in the database — every traversal re-resolves.

    ### Contract invariants
    - Non-2xx responses use the error envelope `{ error: { code, message, details? } }`.
    - Collections return ascending `id` order. Related entities in `to_many`
      traversal sort by target `id` ascending.
    - `POST /resolve` **always** returns 200. Per-path failures appear in
      `response.errors[path]`, never in an envelope.
    - Archived related entities are treated as missing during traversal.
      Archived **root** taxonomies and entities are still served on direct
      lookup — this is intentional.
    - Attributes are scalar only (string / integer / float / boolean /
      date / datetime / null). Objects and arrays are rejected.

...
  contact:
...
servers:
...
paths:
...
  /entities:
    get:
      tags: [entities]
      summary: List entities for a taxonomy.
      description: |
        Scoped by required `taxonomy_id`. Returns in ascending `id` order.
        Archived entities are excluded unless `include_archived=true`.
      operationId: listEntities
      parameters:
        - name: taxonomy_id
          in: query
          required: true
          description: Taxonomy whose entities to list.
          schema: { type: string, minLength: 1 }
        - $ref: "#/components/parameters/IncludeArchived"
      responses:
        "200":
          description: Entity collection.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntityList" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    post:
      tags: [entities]
      summary: Create an entity.
      description: |
        Validates that the referenced taxonomy exists, that attribute keys
        are declared fields, that values match the declared types, and that
        all `required` fields are present. `archived` defaults to `false`.

        Returns `409 conflict` when an entity with the same `id` already
        exists.
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateEntity" }
            examples:
              patient: { $ref: "#/components/examples/PatientEntity" }
      responses:
        "201":
          description: Entity created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /entities/{entity_id}:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    get:
      tags: [entities]
      summary: Get an entity by id.
      description: |
        Archived entities are returned on direct lookup (archive status only
        affects list endpoints and traversal).
      operationId: getEntity
      responses:
        "200":
          description: Entity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    patch:
      tags: [entities]
      summary: Update an entity.
      description: |
        Allowed top-level keys: `archived`, `attributes`. Any other key
        (including `id` and `taxonomy_id`) returns `400 validation_error`.

        `attributes` is **merged** into the existing attribute map; keys
        explicitly set to `null` unset that attribute (legal for
        non-required fields only).
      operationId: patchEntity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PatchEntity" }
            examples:
              setField:
                value:
                  attributes:
                    first_name: Jane
              archive:
                value:
                  archived: true
      responses:
        "200":
          description: Entity updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Entity" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }

    delete:
      tags: [entities]
      summary: Delete an entity.
      description: Hard delete. Returns 204 on success.
      operationId: deleteEntity
      responses:
        "204":
          description: Entity deleted.
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
  /entities/{entity_id}/data:
    parameters:
      - $ref: "#/components/parameters/EntityId"
    get:
      tags: [entities]
      summary: Traverse an entity's related data.
      description: |
        Walks the entity's relationships up to `depth` hops and returns the
        resolved graph. Always includes the root entity's `id` and its
        attributes.

        ### Traversal rules
        - `to_one` with no active match → `null` (nested format).
        - `to_one` with multiple matches → **`409 conflict`** (`ambiguous_to_one`).
        - `to_many` / `to_many_through` are **omitted** unless
          `include_to_many=true`; when included they serialize as arrays
          (nested) or numeric dot-notation (`support_tickets.0.status`) in
          flat format.
        - Archived related entities are treated as missing.

        `format=flat` produces a single-level map with dot-notation keys;
        `format=nested` (default) mirrors the relationship tree.
      operationId: getEntityData
      parameters:
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 2
        - name: include_to_many
          in: query
          required: false
          description: Include `to_many` and `to_many_through` relationships.
          schema:
            type: boolean
            default: false
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [nested, flat]
            default: nested
      responses:
        "200":
          description: Traversed data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntityDataResponse" }
              examples:
                nested:
                  summary: format=nested
                  value:
                    entity_id: patient-1001
                    taxonomy_id: patients
                    data:
                      id: patient-1001
                      patient_number: P1001
                      first_name: John
                      care_team:
                        id: care-team-red
                        team_code: TEAM-RED
                        assigned_nurse: Nurse Joy
                flat:
                  summary: format=flat
                  value:
                    entity_id: patient-1001
                    taxonomy_id: patients
                    data:
                      id: patient-1001
                      patient_number: P1001
                      care_team.id: care-team-red
                      care_team.assigned_nurse: Nurse Joy
                      support_tickets.0.id: ticket-1
                      support_tickets.0.status: open
        "400": { $ref: "#/components/responses/ValidationError" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  parameters:
...
    EntityId:
      name: entity_id
      in: path
      required: true
      schema: { type: string, minLength: 1 }
      description: Entity identifier.
    IncludeArchived:
      name: include_archived
      in: query
      required: false
      description: When `true`, include archived records. Defaults to `false`.
      schema:
        type: boolean
        default: false
  responses:
    ValidationError:
      description: Input failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            field:
              value:
                error:
                  code: validation_error
                  message: "attributes.date_of_birth: expected date (YYYY-MM-DD)"
                  details:
                    fields:
                      "attributes.date_of_birth": expected date
    NotFound:
      description: Resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            entity:
              value:
                error:
                  code: not_found
                  message: "Entity 'patient-9999' not found"
                  details:
                    resource: Entity
                    id: patient-9999
    Conflict:
      description: |
        Request is well-formed but the current state forbids it. Examples:
        duplicate id, taxonomy deletion blocked by entities or inbound refs,
        PATCH taxonomy incompatible with existing entities, or `ambiguous_to_one`
        during `/data` traversal.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            duplicate:
              value:
                error:
                  code: conflict
                  message: "Taxonomy 'patients' already exists"
                  details:
                    id: patients
            incompatible:
              value:
                error:
                  code: conflict
                  message: "field change would orphan existing entity values"
                  details:
                    entities:
                      "patient-1001":
                        - removed field 'legacy_code' still in use
...
    InternalError:
      description: Unhandled server error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
          examples:
            internal:
              value:
                error:
                  code: internal_error
                  message: "internal server error"
  examples:
...
    PatientEntity:
      value:
        id: patient-1001
        taxonomy_id: patients
        archived: false
        attributes:
          patient_number: P1001
          first_name: John
          last_name: Doe
          date_of_birth: "1980-01-15"
          care_team_code: TEAM-RED
  schemas:
...
    ErrorCode:
      type: string
      description: |
        Stable `code` values emitted on non-2xx envelopes. Clients may
        branch on these.
      enum:
        - validation_error
        - not_found
        - conflict
        - response_too_large
        - not_implemented
        - internal_error
        - unavailable
...
    AttributeValue:
      description: |
        A scalar attribute value. Objects and arrays are rejected by the
        validator; `null` is legal on non-required fields.
      oneOf:
        - type: string
        - type: number
        - type: boolean
        - type: "null"

    # ----- Error envelope ----------------------------------------------------
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              description: Optional structured detail about the failure.

    # ----- Taxonomy ----------------------------------------------------------
...
    Entity:
      type: object
      required: [id, taxonomy_id, archived, attributes]
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1 }
        taxonomy_id: { type: string, minLength: 1 }
        archived: { type: boolean }
        attributes:
          type: object
          additionalProperties: { $ref: "#/components/schemas/AttributeValue" }
          description: Flat map of scalar attribute values keyed by field key.
    CreateEntity:
      type: object
      required: [id, taxonomy_id, attributes]
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1 }
        taxonomy_id: { type: string, minLength: 1 }
        archived:
          type: boolean
          default: false
        attributes:
          type: object
          additionalProperties: { $ref: "#/components/schemas/AttributeValue" }
    PatchEntity:
      type: object
      additionalProperties: false
      description: |
        Only `archived` and `attributes` are allowed. `attributes` is MERGED
        into the existing map (not replaced); keys set to `null` unset that
        attribute (legal for non-required fields only). `id` and
        `taxonomy_id` are immutable.
      properties:
        archived: { type: boolean }
        attributes:
          type: object
          additionalProperties: { $ref: "#/components/schemas/AttributeValue" }
    EntityList:
      type: object
      required: [data]
      additionalProperties: false
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Entity" }

    # ----- Health ------------------------------------------------------------
...
    EntityDataResponse:
      type: object
      required: [entity_id, taxonomy_id, data]
      additionalProperties: false
      description: |
        `data` is either a nested object (format=nested) or a flat map with
        dot-notation keys (format=flat). Shape depends on the taxonomy and
        traversal parameters.
      properties:
        entity_id: { type: string }
        taxonomy_id: { type: string }
        data:
          type: object
          additionalProperties: true

    # ----- Resolve -----------------------------------------------------------
...